backend/requirements.txt
fastapi>=0.96.0
uvicorn>=0.22.0
pydantic>=2.0.0
langchain>=0.0.267
langchain-core>=0.1.0
langchain-openai>=0.0.1
langgraph>=0.0.10
python-dotenv>=1.0.0
openai>=1.0.0
